+++ features/table/immutable/src/desktopMain/kotlin/ru/pavlig43/immutable/internal/component/items/product/ProductTableUi.kt
@@ -11,5 +11,5 @@ data class ProductTableUi(
     val type: ProductType,
     val createdAt: LocalDate,
     val comment: String = "",
-    val priceForSale: Int = 0,
+    val priceForSale: Long = 0,
 ) : IMultiLineTableUi
